requirements.txt
cassandra-driver
mmh3
